Phoenix v0.1.2

+++ core/database.cpp
@@ -69,8 +69,9 @@ int Database::insertModel(const QString &name, const QString &key){
     query.addBindValue(key);
     query.addBindValue(QDateTime::currentDateTime());
     query.addBindValue(false);
-    if (!query.exec())
+    if (!query.exec()){
         return -1;
+    }
 
     return query.lastInsertId().toInt();
 }
@@ -115,7 +116,7 @@ void Database::addModel(const QString &name, const QString &key){
 
         emit addOfflineModel(nullptr, fileSize, ramRequ, "", "", "- bilion", "q4_0",0.0, false, true,
                              id, name, name, key, addDate, isLike, "Text Generation", BackendType::OfflineModel,
-                             icon, information, "","", QDateTime::currentDateTime(), false, "");
+                             "qrc:/media/image_company/"+icon, information, "","", QDateTime::currentDateTime(), false, "");
 
     }
 }
@@ -258,7 +259,7 @@ void Database::addHuggingfaceModel(const QString &name, const QString &url, cons
             isLike,                                // isLike
             newModel["type"].toString(),           // type
             BackendType::OfflineModel,             // backend
-            companyIcon,                           // icon
+            "qrc:/media/image_company/"+companyIcon,                           // icon
             information,                           // information
             newModel["promptTemplate"].toString(), // promptTemplate
             newModel["systemPrompt"].toString(),   // systemPrompt
@@ -705,7 +706,7 @@ void Database::readModel(const QList<Company*> companys){
 
                                    id, obj["modelName"].toString(), name, key, addDate, isLike,
                                    obj["type"].toString(), BackendType::OfflineModel,
-                                   company->icon(), obj["description"].toString(), obj["promptTemplate"].toString(),
+                                   "qrc:/media/image_company/"+company->icon(), obj["description"].toString(), obj["promptTemplate"].toString(),
                                    obj["systemPrompt"].toString(), QDateTime::currentDateTime(), obj["recommended"].toBool(), "");
 
                 allID.append(id);
@@ -739,11 +740,11 @@ void Database::readModel(const QList<Company*> companys){
                 bool isLike = query.value(4).toBool();
 
                 QFile file(key);
-                if (!file.exists()){
+                if (!file.exists() && (name !="Indox Router")){
                     deleteModel(id);
-                }else{
+                }else if(name !="Indox Router"){
                     QFileInfo fileInfo(key);
-                    QString icon = "Phoenix.svg";
+                    QString icon = "qrc:/media/image_company/Phoenix.svg";
                     QString information = "This model has been successfully added to the application by you.";
                     double fileSize = (fileInfo.size()/10000000)*0.01;
                     int ramRequ;
@@ -828,9 +829,84 @@ void Database::readMessages(const int idConversation){
 }
 
 QList<int> Database::readOnlineCompany() {
-    QList<int> allID;
+    int indoxRouterId;
+    QString indoxRouterName = "Indox Router";
+    QString indoxRouterKey = "";
+    QDateTime indoxRouterAddDate = QDateTime::currentDateTime();
+    bool indoxRouterIsLike = false;
 
-    QFile file(QCoreApplication::applicationDirPath() + "/models/company.json");
+    QSqlQuery query(m_db);
+    query.prepare(READ_MODEL_SQL);
+    query.addBindValue(indoxRouterName);
+
+    if (query.exec()){
+        if (!query.next()) {
+            indoxRouterId = insertModel(indoxRouterName, indoxRouterKey);
+        } else {
+            indoxRouterId = query.value(0).toInt();
+            indoxRouterName = query.value(1).toString();
+            indoxRouterKey = query.value(2).toString();
+            indoxRouterAddDate = query.value(3).toDateTime();
+            indoxRouterIsLike = query.value(4).toBool();
+        }
+    }
+
+
+    emit addOnlineProvider(indoxRouterId, indoxRouterName,
+                           "qrc:/media/image_company/indoxRoter.png",
+                           indoxRouterIsLike, BackendType::OnlineModel, "",
+                           indoxRouterKey);
+
+    QList<int> allID;
+    QString filePath = QCoreApplication::applicationDirPath() + "/models/online_models/online_models.json";
+
+    bool fileExists = QFile::exists(filePath);
+
+    if (fileExists) {
+
+        QNetworkAccessManager *manager = new QNetworkAccessManager(this);
+        QNetworkRequest request(QUrl("https://api.indoxrouter.com/api/v1/models/"));
+        QNetworkReply *reply = manager->get(request);
+
+        QObject::connect(reply, &QNetworkReply::finished, [reply, filePath]() {
+            if (reply->error() == QNetworkReply::NoError) {
+                QByteArray data = reply->readAll();
+                QDir().mkpath(QFileInfo(filePath).absolutePath());
+                QFile file(filePath);
+                if (file.open(QIODevice::WriteOnly)) {
+                    file.write(data);
+                    file.close();
+                }
+            } else {
+                qWarning() << "Cannot fetch updated online models:" << reply->errorString();
+            }
+            reply->deleteLater();
+        });
+    } else {
+        QNetworkAccessManager manager;
+        QNetworkRequest request(QUrl("https://api.indoxrouter.com/api/v1/models/"));
+        QNetworkReply *reply = manager.get(request);
+
+        QEventLoop loop;
+        QObject::connect(reply, &QNetworkReply::finished, &loop, &QEventLoop::quit);
+        loop.exec();
+
+        if (reply->error() == QNetworkReply::NoError) {
+            QByteArray data = reply->readAll();
+            QDir().mkpath(QFileInfo(filePath).absolutePath());
+            QFile file(filePath);
+            if (file.open(QIODevice::WriteOnly)) {
+                file.write(data);
+                file.close();
+                qInfo() << "Downloaded online models file for the first time.";
+            }
+        } else {
+            qWarning() << "Cannot fetch online models and no cache found!";
+        }
+        reply->deleteLater();
+    }
+
+    QFile file(filePath);
     if (!file.open(QIODevice::ReadOnly)) {
         qWarning() << "Cannot open JSON file!";
         return allID;
@@ -838,53 +914,308 @@ QList<int> Database::readOnlineCompany() {
 
     QByteArray jsonData = file.readAll();
     file.close();
+
     QJsonDocument doc = QJsonDocument::fromJson(jsonData);
     if (!doc.isArray()) {
         qWarning() << "Invalid JSON format!";
         return allID;
     }
 
-    int i=0;
     QJsonArray jsonArray = doc.array();
-    for (const QJsonValue &value : jsonArray){
-        if (!value.isObject()) continue;
+    for (const QJsonValue &value : jsonArray) {
+        if (!value.isObject())
+            continue;
         QJsonObject obj = value.toObject();
 
-        if (obj["type"].toString() == "OnlineModel"){
-            int id;
-            QString name = obj["name"].toString();
-            QString key = "";
-            QDateTime addDate = QDateTime::currentDateTime();
-            bool isLike = false;
-
-            QSqlQuery query(m_db);
-            query.prepare(READ_MODEL_SQL);
-            query.addBindValue(name);
-
-            if (!query.exec())
-                continue;
-
-            bool installModel = false;
-
-            if (!query.next()) {
-                id = insertModel(name, key);
-            }else{
-                id = query.value(0).toInt();
-                name = query.value(1).toString();
-                key = query.value(2).toString();
-                addDate = query.value(3).toDateTime();
-                isLike = query.value(4).toBool();
-                if(key != "")
-                    installModel = true;
-            }
-
-            if(id == -1)
-                continue;
-
-            emit addOnlineProvider(id, name, obj["icon"].toString(), isLike, BackendType::OnlineModel, obj["file"].toString(), key);
-
-            allID.append(id);
+        QJsonArray modelsArray = obj["text_completions"].toArray();
+        if (modelsArray.isEmpty()) {
+            continue;
         }
+
+        int id;
+        QString name = obj["name"].toString();
+        QString key = "";
+        QDateTime addDate = QDateTime::currentDateTime();
+        bool isLike = false;
+
+        QSqlQuery query(m_db);
+        query.prepare(READ_MODEL_SQL);
+        query.addBindValue(name);
+        if (!query.exec())
+            continue;
+
+        if (!query.next()) {
+            id = insertModel(name, key);
+        } else {
+            id = query.value(0).toInt();
+            name = query.value(1).toString();
+            key = query.value(2).toString();
+            addDate = query.value(3).toDateTime();
+            isLike = query.value(4).toBool();
+        }
+
+        if (id == -1)
+            continue;
+
+        QString icon = "";
+        QString idLower = name.toLower();
+
+        static const QVector<QPair<QString, QString>> icons = {
+            {"adobe", "adobe.svg"},
+            {"adobefirefly", "adobefirefly.svg"},
+            {"agui", "agui.svg"},
+            {"ai21", "ai21.svg"},
+            {"ai302", "ai302.svg"},
+            {"ai360", "ai360.svg"},
+            {"aihubmix", "aihubmix.svg"},
+            {"aimass", "aimass.svg"},
+            {"aionlabs", "aionlabs.svg"},
+            {"aistudio", "aistudio.svg"},
+            {"alephalpha", "alephalpha.svg"},
+            {"alibaba", "alibaba.svg"},
+            {"alibabacloud", "alibabacloud.svg"},
+            {"antgroup", "antgroup.svg"},
+            {"anthropic", "anthropic.svg"},
+            {"aya", "aya.svg"},
+            {"azure", "azure.svg"},
+            {"azureai", "azureai.svg"},
+            {"baai", "baai.svg"},
+            {"baichuan", "baichuan.svg"},
+            {"baidu", "baidu.svg"},
+            {"baiducloud", "baiducloud.svg"},
+            {"baseten", "baseten.svg"},
+            {"bedrock", "bedrock.svg"},
+            {"bert", "Bert.svg"},
+            {"bfl", "bfl.svg"},
+            {"bigcode", "BigCode.png"},
+            {"bilibili", "bilibili.svg"},
+            {"bilibiliindex", "bilibiliindex.svg"},
+            {"bing", "bing.svg"},
+            {"burncloud", "burncloud.svg"},
+            {"bytedance", "bytedance.svg"},
+            {"centml", "centml.svg"},
+            {"cerebras", "cerebras.svg"},
+            {"chatglm", "chatglm.svg"},
+            {"civitai", "civitai.svg"},
+            {"claude", "Claude.svg"},
+            {"cline", "cline.svg"},
+            {"clipdrop", "clipdrop.svg"},
+            {"cloudflare", "cloudflare.svg"},
+            {"codegeex", "codegeex.svg"},
+            {"cogvideo", "cogvideo.svg"},
+            {"cogview", "cogview.svg"},
+            {"cohere", "Cohere.svg"},
+            {"colab", "colab.svg"},
+            {"comfyui", "comfyui.svg"},
+            {"commanda", "commanda.svg"},
+            {"copilot", "copilot.svg"},
+            {"copilotkit", "copilotkit.svg"},
+            {"coqui", "coqui.svg"},
+            {"coze", "coze.svg"},
+            {"crewai", "crewai.svg"},
+            {"crusoe", "crusoe.svg"},
+            {"cursor", "cursor.svg"},
+            {"dalle", "dalle.svg"},
+            {"databricks", "Databricks.svg"},
+            {"dbrx", "dbrx.svg"},
+            {"deepai", "deepai.svg"},
+            {"deepinfra", "deepinfra.svg"},
+            {"deepmind", "deepmind.svg"},
+            {"deepseek", "Deepseek.svg"},
+            {"dify", "dify.svg"},
+            {"doc2x", "doc2x.svg"},
+            {"docsearch", "docsearch.svg"},
+            {"dolphin", "dolphin.svg"},
+            {"doubao", "doubao.svg"},
+            {"dreammachine", "dreammachine.svg"},
+            {"elevenlabs", "elevenlabs.svg"},
+            {"elevenx", "elevenx.svg"},
+            {"exa-color", "exa-color.svg"},
+            {"fireworks", "fireworks.svg"},
+            {"flowith", "flowith.svg"},
+            {"flux", "flux.svg"},
+            {"friendli", "friendli.svg"},
+            {"gemini", "gemini.svg"},
+            {"gemma", "gemma.svg"},
+            {"giteeai", "giteeai.svg"},
+            {"githubcopilot", "githubcopilot.svg"},
+            {"glama", "glama.svg"},
+            {"glif", "glif.svg"},
+            {"glmv", "glmv.svg"},
+            {"google", "Google.svg"},
+            {"goose", "goose.svg"},
+            {"gradio", "gradio.svg"},
+            {"greptile", "greptile.svg"},
+            {"grok", "grok.svg"},
+            {"groq", "groq.svg"},
+            {"hailuo", "hailuo.svg"},
+            {"haiper", "haiper.svg"},
+            {"hedra", "hedra.svg"},
+            {"higress", "higress.svg"},
+            {"huggingface", "Huggingface.svg"},
+            {"hunyuan", "hunyuan.svg"},
+            {"hyperbolic", "hyperbolic.svg"},
+            {"ibm", "ibm.svg"},
+            {"ideogram", "ideogram.svg"},
+            {"iflytekcloud", "iflytekcloud.svg"},
+            {"inference", "inference.svg"},
+            {"infermatic", "infermatic.svg"},
+            {"infinigence", "infinigence.svg"},
+            {"inflection", "inflection.svg"},
+            {"internlm", "internlm.svg"},
+            {"jimeng", "jimeng.svg"},
+            {"jina", "jina.svg"},
+            {"kera", "kera.svg"},
+            {"kimi", "kimi.svg"},
+            {"kling", "kling.svg"},
+            {"kluster", "kluster.svg"},
+            {"kolors", "kolors.svg"},
+            {"lambda", "lambda.svg"},
+            {"langchain", "langchain.svg"},
+            {"langfuse", "langfuse.svg"},
+            {"langgraph", "langgraph.svg"},
+            {"langsmith", "langsmith.svg"},
+            {"lepton", "lepton.svg"},
+            {"lg", "lg.svg"},
+            {"lightricks", "lightricks.svg"},
+            {"liquid", "liquid.svg"},
+            {"livekit", "livekit.svg"},
+            {"llamaindex", "llamaindex.svg"},
+            {"llava", "llava.svg"},
+            {"lmstudio", "lmstudio.svg"},
+            {"lobehub", "lobehub.svg"},
+            {"lovable", "lovable.svg"},
+            {"luma", "luma.svg"},
+            {"magic", "magic.svg"},
+            {"make", "make.svg"},
+            {"manus", "manus.svg"},
+            {"mastra", "mastra.svg"},
+            {"mcp", "mcp.svg"},
+            {"mcpso", "mcpso.svg"},
+            {"menlo", "menlo.svg"},
+            {"meta", "Meta.svg"},
+            {"metaai", "metaai.svg"},
+            {"metagpt", "metagpt.svg"},
+            {"microsoft", "Microsoft.svg"},
+            {"midjourney", "midjourney.svg"},
+            {"minimax", "minimax.svg"},
+            {"mistral", "Mistral.svg"},
+            {"modelscope", "modelscope.svg"},
+            {"monica", "monica.svg"},
+            {"moonshot", "moonshot.svg"},
+            {"mpt", "MPT.svg"},
+            {"myshell", "myshell.svg"},
+            {"n8n", "n8n.svg"},
+            {"nebius", "nebius.svg"},
+            {"notebooklm", "notebooklm.svg"},
+            {"notion", "notion.svg"},
+            {"nousresearch", "nousresearch.svg"},
+            {"nova", "nova.svg"},
+            {"novelai", "novelai.svg"},
+            {"novita", "novita.svg"},
+            {"nplcloud", "nplcloud.svg"},
+            {"nvidia", "Nvidia.svg"},
+            {"ollama", "ollama.svg"},
+            {"openai", "Openai.svg"},
+            {"openchat", "openchat.svg"},
+            {"openrouter", "openrouter.svg"},
+            {"openwebui", "openwebui.svg"},
+            {"palm", "palm.svg"},
+            {"parasail", "parasail.svg"},
+            {"perplexity", "perplexity.svg"},
+            {"phidata", "phidata.svg"},
+            {"phind", "phind.svg"},
+            {"phoenix", "Phoenix.svg"},
+            {"pika", "pika.svg"},
+            {"pixverse", "pixverse.svg"},
+            {"player2", "player2.svg"},
+            {"poe", "poe.svg"},
+            {"pollinations", "pollinations.svg"},
+            {"ppio", "ppio.svg"},
+            {"pydanticai", "pydanticai.svg"},
+            {"qingyan", "qingyan.svg"},
+            {"qiniu", "qiniu.svg"},
+            {"qwen", "Qwen.svg"},
+            {"railway", "railway.svg"},
+            {"recraft", "recraft.svg"},
+            {"replicate", "replicate.svg"},
+            {"replit", "Replit.svg"},
+            {"rsshub", "rsshub.svg"},
+            {"runway", "runway.svg"},
+            {"rwkv", "rwkv.svg"},
+            {"sambanova", "sambanova.svg"},
+            {"search1api", "search1api.svg"},
+            {"searchapi", "searchapi.svg"},
+            {"sensenova", "sensenova.svg"},
+            {"siliconcloud", "siliconcloud.svg"},
+            {"skywork", "skywork.svg"},
+            {"smithery", "smithery.svg"},
+            {"snowflake", "snowflake.svg"},
+            {"spark", "spark.svg"},
+            {"stability", "stability.svg"},
+            {"starcoder", "Starcoder.svg"},
+            {"statecloud", "statecloud.svg"},
+            {"stepfun", "stepfun.svg"},
+            {"suno", "suno.svg"},
+            {"sync", "sync.svg"},
+            {"targon", "targon.svg"},
+            {"tavily", "tavily.svg"},
+            {"tencent", "tencent.svg"},
+            {"tencentcloud", "tencentcloud.svg"},
+            {"tiangong", "tiangong.svg"},
+            {"tii", "tii.svg"},
+            {"together", "together.svg"},
+            {"topazlabs", "topazlabs.svg"},
+            {"trae", "trae.svg"},
+            {"tripo", "tripo.svg"},
+            {"udio", "udio.svg"},
+            {"unstructured", "unstructured.svg"},
+            {"upstage", "upstage.svg"},
+            {"user", "user.svg"},
+            {"v0", "v0.svg"},
+            {"vectorizerai", "vectorizerai.svg"},
+            {"vercel", "vercel.svg"},
+            {"vertexai", "vertexai.svg"},
+            {"vidu", "vidu.svg"},
+            {"viggle", "viggle.svg"},
+            {"vllm", "vllm.svg"},
+            {"volcengine", "volcengine.svg"},
+            {"voyage", "voyage.svg"},
+            {"wenxin", "wenxin.svg"},
+            {"whisper", "Whisper.svg"},
+            {"windsurf", "windsurf.svg"},
+            {"workersai", "workersai.svg"},
+            {"xai", "xai.svg"},
+            {"xinference", "xinference.svg"},
+            {"xuanyuan", "xuanyuan.svg"},
+            {"yandex", "yandex.svg"},
+            {"yi", "yi.svg"},
+            {"youmind", "youmind.svg"},
+            {"yuanbao", "yuanbao.svg"},
+            {"zai", "zai.svg"},
+            {"zapier", "zapier.svg"},
+            {"zeabur", "zeabur.svg"},
+            {"zeroone", "zeroone.svg"},
+            {"zhipu", "zhipu.svg"}
+        };
+
+        bool found = false;
+        for (const auto &entry : icons) {
+            if (idLower.contains(entry.first)) {
+                icon = QString("qrc:/media/image_company/%1").arg(entry.second);
+                found = true;
+                break;
+            }
+        }
+
+        if (!found)
+            icon = "qrc:/media/image_company/Huggingface.svg";
+
+        emit addOnlineProvider(id, name, icon, isLike, BackendType::OnlineModel,
+                               obj["file"].toString(), key);
+
+        allID.append(id);
     }
     return allID;
 }
+
